requirements.txt
dbt-postgres==1.3.1
pandas==1.5.2
psycopg2==2.9.5
python-dotenv==0.21.0